refactor: extract common Operation handlers logic to AbstractOperationHandler
